requirements.txt
# ProfMail 依存関係

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Gmail API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# OpenAI API
openai==1.3.5

# Scheduler
APScheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Optional: Development tools
pytest==7.4.3
black==23.11.0
flake8==6.1.0